refactor(prompt): move location permission entry to assistant memory
