+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -417,12 +417,12 @@
   <string name="setting_mcp_page_name_desc">MCP伺服器的顯示名稱</string>
   <string name="setting_mcp_page_name_placeholder">例如：My MCP Server</string>
   <string name="setting_mcp_page_no_mcp_servers_found">找不到 MCP 伺服器</string>
-  <string name="setting_mcp_page_prompt_location_action">授權定位權限</string>
-  <string name="setting_mcp_page_prompt_location_desc">為 MCP 工具暴露提示詞定位變數\n系統提示詞：{{location}}\n訊息模板：{{ location }}\n輸出：lng,lat</string>
-  <string name="setting_mcp_page_prompt_location_status_approximate">狀態：已授權 近似定位</string>
-  <string name="setting_mcp_page_prompt_location_status_denied">狀態：未授權</string>
-  <string name="setting_mcp_page_prompt_location_status_precise">狀態：已授權 精確定位</string>
-  <string name="setting_mcp_page_prompt_location_title">提示詞定位變數</string>
+  <string name="assistant_page_prompt_location_action">授權定位權限</string>
+  <string name="assistant_page_prompt_location_desc">為提示詞變數提供定位座標\n系統提示詞：{{location}}\n訊息模板：{{ location }}\n輸出：lng,lat</string>
+  <string name="assistant_page_prompt_location_status_approximate">狀態：已授權 近似定位</string>
+  <string name="assistant_page_prompt_location_status_denied">狀態：未授權</string>
+  <string name="assistant_page_prompt_location_status_precise">狀態：已授權 精確定位</string>
+  <string name="assistant_page_prompt_location_title">提示詞定位變數</string>
   <string name="setting_mcp_page_save">保存</string>
   <string name="setting_mcp_page_server_url">伺服器地址</string>
   <string name="setting_mcp_page_sse_url_desc">SSE傳輸伺服器的URL地址</string>
