+++ app/src/main/res/values-zh/strings.xml
@@ -419,12 +419,12 @@
   <string name="setting_mcp_page_name_desc">MCP服务器的显示名称</string>
   <string name="setting_mcp_page_name_placeholder">例如：My MCP Server</string>
   <string name="setting_mcp_page_no_mcp_servers_found">未找到 MCP 服务器</string>
-  <string name="setting_mcp_page_prompt_location_action">授权定位权限</string>
-  <string name="setting_mcp_page_prompt_location_desc">为 MCP 工具暴露提示词定位变量\n系统提示词：{{location}}\n消息模板：{{ location }}\n输出：lng,lat</string>
-  <string name="setting_mcp_page_prompt_location_status_approximate">状态：已授权 近似定位</string>
-  <string name="setting_mcp_page_prompt_location_status_denied">状态：未授权</string>
-  <string name="setting_mcp_page_prompt_location_status_precise">状态：已授权 精确定位</string>
-  <string name="setting_mcp_page_prompt_location_title">提示词定位变量</string>
+  <string name="assistant_page_prompt_location_action">授权定位权限</string>
+  <string name="assistant_page_prompt_location_desc">为提示词变量提供定位坐标\n系统提示词：{{location}}\n消息模板：{{ location }}\n输出：lng,lat</string>
+  <string name="assistant_page_prompt_location_status_approximate">状态：已授权 近似定位</string>
+  <string name="assistant_page_prompt_location_status_denied">状态：未授权</string>
+  <string name="assistant_page_prompt_location_status_precise">状态：已授权 精确定位</string>
+  <string name="assistant_page_prompt_location_title">提示词定位变量</string>
   <string name="setting_mcp_page_save">保存</string>
   <string name="setting_mcp_page_server_url">服务器地址</string>
   <string name="setting_mcp_page_sse_url_desc">SSE传输服务器的URL地址</string>
