+++ app/src/main/res/values/strings.xml
@@ -431,12 +431,12 @@
   <string name="setting_mcp_page_name_desc">Display name for the MCP server</string>
   <string name="setting_mcp_page_name_placeholder">e.g.: My MCP Server</string>
   <string name="setting_mcp_page_no_mcp_servers_found">No MCP servers found</string>
-  <string name="setting_mcp_page_prompt_location_action">Grant Permission</string>
-  <string name="setting_mcp_page_prompt_location_desc">Expose location variables in prompts for MCP tools\nSystem prompt: {{location}}\nMessage template: {{ location }}\nOutput: lng,lat</string>
-  <string name="setting_mcp_page_prompt_location_status_approximate">Status: Granted (Approximate)</string>
-  <string name="setting_mcp_page_prompt_location_status_denied">Status: Not granted</string>
-  <string name="setting_mcp_page_prompt_location_status_precise">Status: Granted (Precise)</string>
-  <string name="setting_mcp_page_prompt_location_title">Prompt Location Variables</string>
+  <string name="assistant_page_prompt_location_action">Grant Permission</string>
+  <string name="assistant_page_prompt_location_desc">Provide location coordinates for prompt variables\nSystem prompt: {{location}}\nMessage template: {{ location }}\nOutput: lng,lat</string>
+  <string name="assistant_page_prompt_location_status_approximate">Status: Granted (Approximate)</string>
+  <string name="assistant_page_prompt_location_status_denied">Status: Not granted</string>
+  <string name="assistant_page_prompt_location_status_precise">Status: Granted (Precise)</string>
+  <string name="assistant_page_prompt_location_title">Prompt Location Variable</string>
   <string name="setting_mcp_page_save">Save</string>
   <string name="setting_mcp_page_server_url">Server URL</string>
   <string name="setting_mcp_page_sse_url_desc">URL address for SSE transport server</string>
